\name{mergecalib_fit}
\alias{print.mergecalib_fit}
\alias{summary.mergecalib_fit}
\alias{print.summary.mergecalib_fit}
\alias{plot.mergecalib_fit}
\title{Methods for mergecalib fits}
\usage{
\method{print}{mergecalib_fit}(x, ...)
\method{summary}{mergecalib_fit}(object, ...)
\method{print}{summary.mergecalib_fit}(x, ...)
\method{plot}{mergecalib_fit}(x, type = c("target_error", "cluster_size"), ...)
}
\arguments{
\item{x}{A fitted object or summary object.}
\item{object}{A fitted object.}
\item{type}{Plot type.}
\item{...}{Additional arguments.}
}
\description{Print, summarize, and plot fitted cell-merging models.}
